20200727 유효성 검사
추가

+++ team_cafe/src/main/java/com/cafe/erp/hr/model/salaryDAO.java
@@ -24,5 +24,5 @@ public interface salaryDAO {
 
 	public int getSalaryCount();
 	
-	public int getMaxMonthTotal(HashMap map);
+	public double getMaxMonthTotal(HashMap map);
 }
